fnvsum; related test code and docs; version bump

+++ include/fnv.h
@@ -9,12 +9,13 @@ See <https://en.wikipedia.org/wiki/Fowler%E2%80%93Noll%E2%80%93Vo_hash_function>
 (c) 2016 Sirio Balmelli and Anthony Soenen; https://b-ad.ch
 */
 
+#include <stddef.h> /* size_t */
 #include <stdint.h>
 #include <nonlibc.h>
 
 
-NLC_PUBLIC uint64_t	fnv_hash64(uint64_t *hash, const void *data, uint64_t data_len);
-NLC_PUBLIC uint32_t	fnv_hash32(uint32_t *hash, const void *data, uint64_t data_len);
+NLC_PUBLIC uint64_t	fnv_hash64(uint64_t *hash, const void *data, size_t data_len);
+NLC_PUBLIC uint32_t	fnv_hash32(uint32_t *hash, const void *data, size_t data_len);
 
 
 #endif /* fnv_h_ */
